requirements.txt
streamlit==1.28.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
xlsxwriter==3.1.9
openpyxl==3.1.2
bcrypt==4.0.1
pillow==10.1.0
statsmodels==0.14.0
scikit-learn==1.3.2
scipy==1.11.4
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0